+++ packages/workbox-google-analytics/src/utils/constants.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import '../_version.mjs';
+import '../_version.js';
 
 export const QUEUE_NAME = 'workbox-google-analytics';
 export const MAX_RETENTION_TIME = 60 * 48; // Two days in minutes
